Add dev command.
