+++ src/commands/deploy.ts
@@ -1,37 +1,27 @@
 import { flags } from '@oclif/command';
-import axios, { AxiosResponse } from 'axios';
 import chalk from 'chalk';
 import cli from 'cli-ux';
-import execa from 'execa';
-import fs from 'fs-extra';
 import inquirer from 'inquirer';
-import isCi from 'is-ci';
-import yaml from 'js-yaml';
-import opener from 'opener';
 import AccountUtils from '../architect/account/account.utils';
 import { EnvironmentUtils } from '../architect/environment/environment.utils';
 import PipelineUtils from '../architect/pipeline/pipeline.utils';
 import Command from '../base-command';
-import LocalDependencyManager from '../common/dependency-manager/local-manager';
-import { DockerComposeUtils } from '../common/docker-compose';
-import DockerComposeTemplate from '../common/docker-compose/template';
 import { DeploymentFailedError, PipelineAbortedError, PollingTimeout } from '../common/errors/pipeline-errors';
-import * as Docker from '../common/utils/docker';
-import { ComponentSlugUtils, ComponentSpec, ComponentVersionSlugUtils } from '../dependency-manager/src';
-import { buildSpecFromPath } from '../dependency-manager/src/spec/utils/component-builder';
+import { getComponentSecrets, getInterfacesMap, parseFlags } from '../common/utils/deploy';
 import { Dictionary } from '../dependency-manager/src/utils/dictionary';
+
 
 export abstract class DeployCommand extends Command {
 
   static flags = {
     ...Command.flags,
     auto_approve: flags.boolean({
-      exclusive: ['local', 'compose-file', 'compose_file'],
+      exclusive: ['compose-file', 'compose_file'],
       description: `${Command.DEPRECATED} Please use --auto-approve.`,
       hidden: true,
     }),
     'auto-approve': flags.boolean({
-      exclusive: ['local', 'compose-file', 'compose_file'],
+      exclusive: ['compose-file', 'compose_file'],
       description: 'Automatically approve the deployment without a review step. Used for debugging and CI flows.',
     }),
   };
@@ -70,42 +60,16 @@ export abstract class DeployCommand extends Command {
 
 export default class Deploy extends DeployCommand {
   auth_required(): boolean {
-    const { flags } = this.parse(Deploy);
-    return !flags.local;
+    return true;
   }
 
-  static description = 'Create a deploy job on Architect Cloud or run stacks locally';
+  static description = 'Create a deploy job on Architect Cloud';
 
   static flags = {
     ...DeployCommand.flags,
     ...AccountUtils.flags,
     ...EnvironmentUtils.flags,
 
-    local: flags.boolean({
-      char: 'l',
-      description: 'Deploy the stack locally instead of via Architect Cloud',
-      exclusive: ['account', 'auto-approve', 'auto_approve', 'refresh'],
-    }),
-    production: flags.boolean({
-      description: 'Build and run components without debug blocks',
-      dependsOn: ['local'],
-    }),
-    compose_file: flags.string({
-      description: `${Command.DEPRECATED} Please use --compose-file.`,
-      exclusive: ['account', 'environment', 'auto-approve', 'auto_approve', 'refresh'],
-      hidden: true,
-    }),
-    'compose-file': flags.string({
-      char: 'o',
-      description: 'Path where the compose file should be written to',
-      default: '',
-      exclusive: ['account', 'environment', 'auto-approve', 'auto_approve', 'refresh'],
-    }),
-    detached: flags.boolean({
-      description: 'Run in detached mode',
-      char: 'd',
-      dependsOn: ['local'],
-    }),
     parameter: flags.string({
       char: 'p',
       description: 'Component parameters',
@@ -131,32 +95,12 @@ export default class Deploy extends DeployCommand {
       default: true,
       allowNo: true,
       description: '[default: true] Toggle for deletion protection on deployments',
-      exclusive: ['local'],
     }),
     recursive: flags.boolean({
       char: 'r',
       default: true,
       allowNo: true,
       description: '[default: true] Toggle to automatically deploy all dependencies',
-    }),
-    refresh: flags.boolean({
-      default: true,
-      hidden: true,
-      allowNo: true,
-      exclusive: ['local', 'compose-file', 'compose_file'],
-    }),
-    browser: flags.boolean({
-      default: true,
-      allowNo: true,
-      description: '[default: true] Automatically open urls in the browser for local deployments',
-    }),
-    build_parallel: flags.boolean({
-      description: `${Command.DEPRECATED} Please use --build-parallel.`,
-      hidden: true,
-    }),
-    'build-parallel': flags.boolean({
-      default: false,
-      description: '[default: false] Build docker images in parallel',
     }),
   };
 
@@ -174,142 +118,8 @@ export default class Deploy extends DeployCommand {
     const parsed = super.parse(options, argv);
     parsed.args.configs_or_components = parsed.argv;
 
-    // Merge any values set via deprecated flags into their supported counterparts
-    const flags: any = parsed.flags;
-    flags['build-parallel'] = flags.build_parallel ? flags.build_parallel : flags['build-parallel'];
-    flags['compose-file'] = flags.compose_file ? flags.compose_file : flags['compose-file'];
-    flags['secrets'] = flags.values ? flags.values : flags['secrets'];
-
-    // If values were provided and secrets were not provided, override the secrets with the values
-    if (!flags.secrets && fs.existsSync('./values.yml')) {
-      flags.secrets = './values.yml';
-    } else if (!flags.secrets && fs.existsSync('./secrets.yml')) {
-      flags.secrets = './secrets.yml';
-    }
-
-    parsed.flags = flags;
-
+    parsed.flags = parseFlags(parsed.flags);
     return parsed;
-  }
-
-  async runCompose(compose: DockerComposeTemplate): Promise<void> {
-    const { flags } = this.parse(Deploy);
-
-    const project_name = flags.environment || DockerComposeUtils.DEFAULT_PROJECT;
-    const compose_file = flags['compose-file'] || DockerComposeUtils.buildComposeFilepath(this.app.config.getConfigDir(), project_name);
-
-    await fs.ensureFile(compose_file);
-    await fs.writeFile(compose_file, yaml.dump(compose));
-    this.log(`Wrote docker-compose file to: ${compose_file}`);
-
-    if (flags['build-parallel']) {
-      await execa('docker-compose', ['-f', compose_file, '-p', project_name, 'build', '--parallel'], { stdio: 'inherit' });
-    } else {
-      await execa('docker-compose', ['-f', compose_file, '-p', project_name, 'build'], { stdio: 'inherit' });
-    }
-
-    console.clear();
-
-    this.log('Building containers...', chalk.green('done'));
-    this.log('');
-
-    this.log('Once the containers are running they will be accessible via the following urls:');
-
-    const exposed_interfaces: string[] = [];
-    const gateway = compose.services['gateway'];
-    if (gateway?.ports?.length && typeof gateway.ports[0] === 'string') {
-      const gateway_port = gateway.ports[0].split(':')[0];
-      for (const [service_name, service] of Object.entries(compose.services)) {
-        if (service.labels?.includes('traefik.enable=true')) {
-          const host_rules = service.labels.filter(label => label.includes('rule=Host'));
-          for (const host_rule of host_rules) {
-            const host = new RegExp(/Host\(`(.*?)`\)/g);
-            const host_match = host.exec(host_rule);
-            if (host_match) {
-              this.log(`${chalk.blue(`http://${host_match[1]}:${gateway_port}/`)} => ${service_name}`);
-              exposed_interfaces.push(`http://${host_match[1]}:${gateway_port}/`);
-            }
-          }
-        }
-      }
-      this.log('');
-    }
-
-    for (const svc_name of Object.keys(compose.services)) {
-      for (const port_pair of compose.services[svc_name].ports || []) {
-        const [exposed_port, internal_port] = port_pair && (port_pair as string).split(':');
-        this.log(`${chalk.blue(`http://localhost:${exposed_port}/`)} => ${svc_name}:${internal_port}`);
-      }
-    }
-    this.log('');
-    this.log('Starting containers...');
-    this.log('');
-
-    if (!isCi && flags.browser) {
-      let open_browser_attempts = 0;
-      const poll_interval = 2000;
-      const browser_interval = setInterval(async () => {
-        if (open_browser_attempts === 150) {
-          clearInterval(browser_interval);
-          return;
-        }
-
-        const promises: Promise<AxiosResponse<any>>[] = [];
-        for (const exposed_interface of exposed_interfaces) {
-          const [host_name, port] = exposed_interface.replace('http://', '').split(':');
-          promises.push(axios.get(`http://localhost:${port}`, {
-            headers: {
-              Host: host_name,
-            },
-            maxRedirects: 0,
-            timeout: poll_interval,
-            validateStatus: (status: number) => { return status < 500 && status !== 404; },
-          }));
-        }
-
-        Promise.all(promises).then(() => {
-          for (const exposed_interface of exposed_interfaces) {
-            this.log('Opening', chalk.blue(exposed_interface));
-            opener(exposed_interface);
-          }
-          this.log('(disable with --no-browser)');
-          clearInterval(browser_interval);
-        }).catch(err => {
-          // at least one exposed service is not yet ready
-        });
-        open_browser_attempts++;
-      }, poll_interval);
-    }
-
-    const compose_args = ['-f', compose_file, '-p', project_name, 'up', '--timeout', '0'];
-    if (flags.detached) {
-      compose_args.push('-d');
-    }
-
-    const cmd = execa('docker-compose', compose_args);
-    cmd.stdin?.pipe(process.stdin);
-    cmd.stdout?.pipe(process.stdout);
-    cmd.stderr?.pipe(process.stderr);
-
-    process.on('SIGINT', () => {
-      this.log('Interrupt received.');
-      this.warn('Please wait for architect to exit or containers will still be running in the background.');
-      this.log('Gracefully shutting down...');
-      execa.sync('docker-compose', ['-f', compose_file, '-p', project_name, 'stop', '--timeout', '0'], { stdio: 'inherit' });
-      this.log('Stopping operation...');
-      process.exit(0);
-    });
-
-    await cmd;
-  }
-
-  private readSecretsFile(secrets_file_path: string | undefined) {
-    let component_secrets: any = {};
-    if (secrets_file_path && fs.statSync(secrets_file_path)) {
-      const secrets_file_data = fs.readFileSync(secrets_file_path);
-      component_secrets = yaml.load(secrets_file_data.toString('utf-8'));
-    }
-    return component_secrets;
   }
 
   getExtraEnvironmentVariables(parameters: string[]): Dictionary<string | number | undefined> {
@@ -341,92 +151,13 @@ export default class Deploy extends DeployCommand {
     return extra_env_vars;
   }
 
-  private getComponentSecrets() {
-    const { flags } = this.parse(Deploy);
-    const component_secrets = this.readSecretsFile(flags.secrets);
-    const extra_params = this.getExtraEnvironmentVariables(flags.parameter);
-    if (extra_params && Object.keys(extra_params).length) {
-      if (!component_secrets['*']) {
-        component_secrets['*'] = {};
-      }
-      component_secrets['*'] = { ...component_secrets['*'], ...extra_params };
-    }
-    return component_secrets;
-  }
-
-  private getInterfacesMap() {
-    const { flags } = this.parse(Deploy);
-    const interfaces_map: Dictionary<string> = {};
-    for (const i of flags.interface) {
-      const [key, value] = i.split(':');
-      interfaces_map[key] = value || key;
-    }
-    return interfaces_map;
-  }
-
-  private async runLocal() {
-    const { args, flags } = this.parse(Deploy);
-    await Docker.verify();
-
-    if (!args.configs_or_components || !args.configs_or_components.length) {
-      args.configs_or_components = ['./architect.yml'];
-    }
-
-    const interfaces_map = this.getInterfacesMap();
-    const component_secrets = this.getComponentSecrets();
-
-    const linked_components = this.app.linkedComponents;
-    const component_versions: string[] = [];
-    for (const config_or_component of args.configs_or_components) {
-
-      let component_version = config_or_component;
-      if (!ComponentVersionSlugUtils.Validator.test(config_or_component) && !ComponentSlugUtils.Validator.test(config_or_component)) {
-        const res = buildSpecFromPath(config_or_component);
-        linked_components[res.name] = config_or_component;
-        component_version = res.name;
-      }
-      component_versions.push(component_version);
-    }
-
-    const dependency_manager = new LocalDependencyManager(
-      this.app.api,
-      linked_components,
-      flags.production
-    );
-
-    const component_specs: ComponentSpec[] = [];
-
-    // Check if multiple instances of the same component are being deployed. This check is needed
-    // so that we can disable automatic interface mapping since we can't map a single interface to
-    // multiple components at this time
-    const onlyUnique = <T>(value: T, index: number, self: T[]) => self.indexOf(value) === index;
-    const uniqe_names = component_versions.map(name => name.split('@')[0]).filter(onlyUnique);
-    const duplicates = uniqe_names.length !== component_versions.length;
-
-    const component_options = { map_all_interfaces: !flags.production && !duplicates };
-
-    for (const component_version of component_versions) {
-      const component_config = await dependency_manager.loadComponentSpec(component_version, interfaces_map, component_options);
-
-      if (flags.recursive) {
-        const dependency_configs = await dependency_manager.loadComponentSpecs(component_config);
-        component_specs.push(...dependency_configs);
-      } else {
-        component_specs.push(component_config);
-      }
-    }
-    const graph = await dependency_manager.getGraph(component_specs, component_secrets);
-    const compose = await DockerComposeUtils.generate(graph);
-    await this.runCompose(compose);
-  }
-
   protected async runRemote(): Promise<void> {
     const { args, flags } = this.parse(Deploy);
 
     const components = args.configs_or_components;
 
-    const interfaces_map = this.getInterfacesMap();
-    const component_secrets = this.getComponentSecrets();
+    const interfaces_map = getInterfacesMap(flags.interface);
+    const component_secrets = getComponentSecrets(flags.secrets, flags.parameter);
 
     const account = await AccountUtils.getAccount(this.app, flags.account);
     const environment = await EnvironmentUtils.getEnvironment(this.app.api, account, flags.environment);
@@ -493,10 +224,6 @@ export default class Deploy extends DeployCommand {
       }
     }
 
-    if (flags.local) {
-      await this.runLocal();
-    } else {
-      await this.runRemote();
-    }
+    await this.runRemote();
   }
 }
